refactor: Token 삭제 시 예외처리 강화
Token이 존재하지 않는 경우 예외를 처리하도록 예외처리를 강화하였습니다

Resolves: #9
Related to: #1

+++ src/main/kotlin/rainbowfriends/daramserverv2/global/security/component/DeleteToken.kt
@@ -1,11 +1,16 @@
 package rainbowfriends.daramserverv2.global.security.component
 
 import org.springframework.stereotype.Component
+import rainbowfriends.daramserverv2.domain.auth.exception.TokenNotFoundException
 import rainbowfriends.daramserverv2.global.redis.RedisUtil
 
 @Component
 class DeleteToken(private val redisUtil: RedisUtil) {
     fun deleteToken(token: String) {
-        redisUtil.delete(token)
+        if (redisUtil.exists(token)) {
+            redisUtil.delete(token)
+        } else {
+            throw TokenNotFoundException("Token Not Found")
+        }
     }
 }
